requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
requests
python-dotenv